Check library controller

+++ src/library/accessor.library.ts
@@ -11,7 +11,11 @@ import { searchLibrary } from './search';
 export class AccessorLibrary {
   constructor(readonly dir: string) {}
 
-  async search(lastTime: number, limit: number, keywords: string): Promise<MangaOutline[]> {
+  async search(
+    lastTime: number | undefined,
+    limit: number,
+    keywords: string
+  ): Promise<MangaOutline[]> {
     return searchLibrary(this.dir, lastTime, limit, keywords);
   }
 
@@ -22,12 +26,14 @@ export class AccessorLibrary {
   }
 
   async createManga(mangaId: string): Promise<void> {
+    if (!this.validateMangaId(mangaId)) return;
     const mangaDir = path.join(this.dir, mangaId);
     if (fsu.isDirectoryExist(mangaDir)) return;
     return fs.mkdir(mangaDir);
   }
 
   async deleteManga(mangaId: string): Promise<void> {
+    if (!this.validateMangaId(mangaId)) return;
     const mangaDir = path.join(this.dir, mangaId);
     if (!fsu.isDirectoryExist(mangaDir)) return;
     return fs.rmdir(mangaDir, { recursive: true });
